fix: fixed semaphore and rate limiter positions and did some refactoring
