--- src/main/kotlin/ru/quipy/common/utils/TooManyRequestsException.kt
@@ -1,3 +1,0 @@
-package ru.quipy.common.utils
-
-class TooManyRequestsException(msg: String) : RuntimeException(msg)
