+++ src/main/kotlin/ru/quipy/exceptions/TooManyRequestsException.kt
@@ -1,0 +1,3 @@
+package ru.quipy.exceptions
+
+class TooManyRequestsException() : RuntimeException()
